extract User model
